fix issue 3

+++ Issue solutions/issue3.cpp
@@ -3,8 +3,8 @@ using namespace std;
 int main() {
     long long int n;
     cin>>n;
-    for (int i = 0 ; i < n ; i++) {
-        for(int j = 0 ; j <= i ; j++) {
+    for (long long int i = 0 ; i < n ; i++) {
+        for(long long int j = 0 ; j <= i ; j++) {
             cout<<"*";
         }
         cout<<endl;
